Added support for the --vm option in corehost.  It both supports --vm as a runtime option and you can read it in via runtimeconfig.json.
It's a bit of hack as it does not muck with the deps_resolver.  That assumes everything
is in the same directory and I didn't want to add a few if's around the mono specific stuff.

You'll need to add libmono.dylib to Microsoft.NETCore.App.deps.json.  Additionally,
you should create a mono directory in shared/<version>.  This is where libmono.dylib
and mono's SPC.dll should live.